cohort/APi Call/gemin/requirements.txt
python-dotenv
openai
google-generativeai

